Removed the checking of libregex
Check libregex makes no good in Windows:
For Windows XP, it's already included in the IUP porting to avoid old gcc issues.
For Windows 32/64, the MSYS2 libregex is not in the native lib, which requires libsystre-0.dll.
So always build the static libregex from source codes in MSYS2.
For Linux, libregex is part of libc, which is always available.